fix: 몬스터 방향 적용 완료

+++ KirbyServer/KirbyServer/Object.cpp
@@ -16,6 +16,6 @@ PAIR Object::NormalizationDir(POINT target)
 	double magnitude = sqrt((target.x - position.x) * (target.x - position.x) + (target.y - position.y) * (target.y - position.y));
 	if (magnitude < 1e-3)
 		return { 0, 0 };
-	dir = { (target.x - position.x) / magnitude, (target.y - position.y) / magnitude };
+	dir = { (double)(target.x - position.x) / magnitude, (double)(target.y - position.y) / magnitude };
 	return dir;
 }
